refactor: 전체 패키지 구조 정리
- - domain 하위는 컨트롤러-서비스-리포지토리 + 관련 DTO와 엔티티를
  하나의 패키지로
- common 하위는 기능별 or 사용 기술별 분리
- test 하위 변경사항  맞추어 변경
