+++ src/main/java/com/concert/ticketing/common/lock/RedisLockRepository.java
@@ -1,4 +1,4 @@
-package com.concert.ticketing.common.repository;
+package com.concert.ticketing.common.lock;
 
 import java.time.Duration;
