+++ src/main/java/com/concert/ticketing/domain/auth/controller/AuthController.java
@@ -1,4 +1,4 @@
-package com.concert.ticketing.domain.member.controller;
+package com.concert.ticketing.domain.auth.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -7,24 +7,21 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.concert.ticketing.common.response.ApiResponse;
-import com.concert.ticketing.domain.member.dto.LoginRequestDto;
-import com.concert.ticketing.domain.member.dto.LoginResponseDto;
-import com.concert.ticketing.domain.member.dto.SignupRequestDto;
-import com.concert.ticketing.domain.member.service.AuthService;
+import com.concert.ticketing.common.dto.ApiResponse;
+import com.concert.ticketing.domain.auth.dto.LoginRequestDto;
+import com.concert.ticketing.domain.auth.dto.LoginResponseDto;
+import com.concert.ticketing.domain.auth.dto.SignupRequestDto;
+import com.concert.ticketing.domain.auth.service.AuthService;
 
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 
 @RequiredArgsConstructor
 @RequestMapping("/api")
 @RestController
-@Slf4j
 public class AuthController {
 
 	private final AuthService authService;
 
-	// 회원가입 - 공동객체로 수정
 	@PostMapping("/signup")
 	public ResponseEntity<ApiResponse<Void>> signup(
 		@RequestBody SignupRequestDto request
@@ -39,7 +36,6 @@ public class AuthController {
 		return ResponseEntity.status(HttpStatus.CREATED).body(response);
 	}
 
-	// 로그인 - 공동객체로 수정
 	@PostMapping("/login")
 	public ResponseEntity<ApiResponse<LoginResponseDto>> login(
 		@RequestBody LoginRequestDto request
