+++ src/main/java/com/concert/ticketing/domain/auth/dto/LoginRequestDto.java
@@ -1,0 +1,11 @@
+package com.concert.ticketing.domain.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class LoginRequestDto {
+	String email;
+	String password;
+}
