+++ src/main/java/com/concert/ticketing/domain/auth/dto/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package com.concert.ticketing.domain.member.dto;
+package com.concert.ticketing.domain.auth.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
