+++ src/main/java/com/concert/ticketing/domain/concert/controller/ConcertController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.concert.ticketing.common.response.ApiResponse;
+import com.concert.ticketing.common.dto.ApiResponse;
 import com.concert.ticketing.domain.concert.dto.ConcertResponse;
 import com.concert.ticketing.domain.concert.service.ConcertService;
