--- src/main/java/com/concert/ticketing/domain/concert/entity/Sector.java
@@ -1,7 +1,0 @@
-package com.concert.ticketing.domain.concert.entity;
-
-public enum Sector {
-	A,
-	B,
-	C
-}
