+++ src/main/java/com/concert/ticketing/domain/seat/controller/RemainSeatController.java
@@ -1,13 +1,13 @@
-package com.concert.ticketing.domain.concert.controller;
+package com.concert.ticketing.domain.seat.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.concert.ticketing.common.response.ApiResponse;
-import com.concert.ticketing.domain.concert.seatDto.RemainSeatResponseDto;
-import com.concert.ticketing.domain.concert.seatService.RemainSeatService;
+import com.concert.ticketing.common.dto.ApiResponse;
+import com.concert.ticketing.domain.seat.dto.RemainSeatResponseDto;
+import com.concert.ticketing.domain.seat.service.RemainSeatService;
 
 import lombok.RequiredArgsConstructor;
