+++ src/main/java/com/concert/ticketing/domain/sector/type/Sector.java
@@ -1,0 +1,7 @@
+package com.concert.ticketing.domain.sector.type;
+
+public enum Sector {
+	A,
+	B,
+	C
+}
