+++ src/main/java/com/concert/ticketing/domain/ticket/repository/TicketCustomRepository.java
@@ -2,7 +2,7 @@ package com.concert.ticketing.domain.ticket.repository;
 
 import java.util.List;
 
-import com.concert.ticketing.domain.ticket.dto.response.TicketResponse;
+import com.concert.ticketing.domain.ticket.dto.TicketResponse;
 
 public interface TicketCustomRepository {
