+++ src/test/java/com/concert/ticketing/domain/ticket/service/TicketServiceTest.java
@@ -19,7 +19,7 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.jdbc.Sql;
 
 import com.concert.ticketing.common.exception.CustomException;
-import com.concert.ticketing.domain.concert.entity.Sector;
+import com.concert.ticketing.domain.sector.type.Sector;
 import com.concert.ticketing.domain.member.entity.Member;
 import com.concert.ticketing.domain.member.repository.MemberRepository;
 import com.concert.ticketing.domain.ticket.repository.TicketRepository;
